+++ src/Controller/ScheduledDataflowController.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace CodeRhapsodie\EzDataflowBundle\Controller;
 
+use CodeRhapsodie\DataflowBundle\Entity\ScheduledDataflow;
 use CodeRhapsodie\EzDataflowBundle\Form\CreateScheduledType;
 use CodeRhapsodie\EzDataflowBundle\Form\UpdateScheduledType;
-use CodeRhapsodie\EzDataflowBundle\Gateway\ScheduledDataflowGateway;
 use CodeRhapsodie\EzDataflowBundle\Gateway\JobGateway;
-use CodeRhapsodie\DataflowBundle\Entity\ScheduledDataflow;
+use CodeRhapsodie\EzDataflowBundle\Gateway\ScheduledDataflowGateway;
 use eZ\Publish\Core\MVC\Symfony\Security\Authorization\Attribute;
 use EzSystems\EzPlatformAdminUi\Notification\NotificationHandlerInterface;
 use EzSystems\EzPlatformAdminUiBundle\Controller\Controller;
@@ -16,7 +16,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Translation\TranslatorInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * @Route("/ezdataflow/scheduled_workflow")
@@ -32,8 +32,12 @@ class ScheduledDataflowController extends Controller
     /** @var TranslatorInterface */
     private $translator;
 
-    public function __construct(JobGateway $jobGateway, NotificationHandlerInterface $notificationHandler, ScheduledDataflowGateway $scheduledDataflowGateway, TranslatorInterface $translator)
-    {
+    public function __construct(
+        JobGateway $jobGateway,
+        NotificationHandlerInterface $notificationHandler,
+        ScheduledDataflowGateway $scheduledDataflowGateway,
+        TranslatorInterface $translator
+    ) {
         $this->jobGateway = $jobGateway;
         $this->notificationHandler = $notificationHandler;
         $this->scheduledDataflowGateway = $scheduledDataflowGateway;
@@ -42,10 +46,6 @@ class ScheduledDataflowController extends Controller
 
     /**
      * @Route("/create", name="coderhapsodie.ezdataflow.workflow.create", methods={"POST"})
-     *
-     * @param Request $request
-     *
-     * @return Response
      */
     public function create(Request $request): Response
     {
@@ -63,7 +63,8 @@ class ScheduledDataflowController extends Controller
                 $this->scheduledDataflowGateway->save($newWorkflow);
                 $this->notificationHandler->success($this->translator->trans('coderhapsodie.ezdataflow.workflow.create.success'));
             } catch (\Exception $e) {
-                $this->notificationHandler->error($this->translator->trans('coderhapsodie.ezdataflow.workflow.create.error', ['message' => $e->getMessage()]));
+                $this->notificationHandler->error($this->translator->trans('coderhapsodie.ezdataflow.workflow.create.error',
+                    ['message' => $e->getMessage()]));
             }
 
             return new JsonResponse(['redirect' => $this->generateUrl('coderhapsodie.ezdataflow.main')]);
@@ -79,10 +80,6 @@ class ScheduledDataflowController extends Controller
 
     /**
      * @Route("/{id}/delete", name="coderhapsodie.ezdataflow.workflow.delete", methods={"post"})
-     *
-     * @param int $id
-     *
-     * @return Response
      */
     public function delete(int $id): Response
     {
@@ -94,7 +91,8 @@ class ScheduledDataflowController extends Controller
 
             return new JsonResponse(['code' => 200]);
         } catch (\Exception $e) {
-            $this->notificationHandler->error($this->translator->trans('coderhapsodie.ezdataflow.workflow.delete.error', ['message' => $e->getMessage()]));
+            $this->notificationHandler->error($this->translator->trans('coderhapsodie.ezdataflow.workflow.delete.error',
+                ['message' => $e->getMessage()]));
 
             return new JsonResponse(['code' => $e->getCode()]);
         }
@@ -102,11 +100,6 @@ class ScheduledDataflowController extends Controller
 
     /**
      * @Route("/{id}/edit", name="coderhapsodie.ezdataflow.workflow.edit")
-     *
-     * @param Request $request
-     * @param int     $id
-     *
-     * @return Response
      */
     public function edit(Request $request, int $id): Response
     {
@@ -123,7 +116,8 @@ class ScheduledDataflowController extends Controller
                 $this->scheduledDataflowGateway->save($editDataflow);
                 $this->notificationHandler->success($this->translator->trans('coderhapsodie.ezdataflow.workflow.edit.success'));
             } catch (\Exception $e) {
-                $this->notificationHandler->error($this->translator->trans('coderhapsodie.ezdataflow.workflow.edit.error', ['message' => $e->getMessage()]));
+                $this->notificationHandler->error($this->translator->trans('coderhapsodie.ezdataflow.workflow.edit.error',
+                    ['message' => $e->getMessage()]));
             }
 
             return new JsonResponse(['redirect' => $this->generateUrl('coderhapsodie.ezdataflow.main')]);
@@ -139,32 +133,12 @@ class ScheduledDataflowController extends Controller
 
     /**
      * @Route("/{id}/enable", name="coderhapsodie.ezdataflow.workflow.enable")
-     *
-     * @param int $id
-     *
-     * @return Response
      */
     public function enableDataflow(int $id): Response
     {
         $this->denyAccessUnlessGranted(new Attribute('ezdataflow', 'edit'));
 
         $this->changeDataflowStatus($id, true);
-
-        return $this->redirectToRoute('coderhapsodie.ezdataflow.main');
-    }
-
-    /**
-     * @Route("/{id}/disable", name="coderhapsodie.ezdataflow.workflow.disable")
-     *
-     * @param int $id
-     *
-     * @return Response
-     */
-    public function disableDataflow(int $id): Response
-    {
-        $this->denyAccessUnlessGranted(new Attribute('ezdataflow', 'edit'));
-
-        $this->changeDataflowStatus($id, false);
 
         return $this->redirectToRoute('coderhapsodie.ezdataflow.main');
     }
@@ -181,5 +155,17 @@ class ScheduledDataflowController extends Controller
         } catch (\Exception $e) {
             $this->notificationHandler->error(sprintf('An error occured : "%s".', $e->getMessage()));
         }
+    }
+
+    /**
+     * @Route("/{id}/disable", name="coderhapsodie.ezdataflow.workflow.disable")
+     */
+    public function disableDataflow(int $id): Response
+    {
+        $this->denyAccessUnlessGranted(new Attribute('ezdataflow', 'edit'));
+
+        $this->changeDataflowStatus($id, false);
+
+        return $this->redirectToRoute('coderhapsodie.ezdataflow.main');
     }
 }